feat(guest): create task/process supervisor and impl for init
This change introduces a Supervisor for async tasks and child processes.
The Supervisor takes a specification that describes how to start, as
well as if and how to retry, said units, and then runs them. This
provides a framework for resilience against ephemeral errors as well as
a standard root for ensuring consistency in how our concurrency
operates.

The Supervisor takes over as the root concept of the guest init.
User Workloads are now just child processes to be spawned. Background
tasks get provided the context they need for coordination inbetween
tasks across processes/domains (via IDM).

In a later change, tasks will be provided the means to spawn their
dependencies in such a way that task tree lifecycles are holistically
maintained by the Supervisor.